// src/types/Movie.ts
export interface IMovie {
    _id: string;
    name: string;
    thumbnail: string;
    author: string;
    actors: string;
    trailer_embed?: string;
    categories: string;
    show_duration: number;
    rated_id: string;
    liked: number;
    description: string;
    language: string;
    createdAt: string;
    updatedAt: string;
    __v: number;
}